* ktest.c (ktest_make_sample_etype_info): Use KRB5_ETYPE_NO_SALT         instead of -1.
        * ktest_equal.c (ktest_equal_krb5_etype_info_entry): Test for
        KRB5_ETYPE_NO_SALT. ktest_equal_array_of_octet() and
        ktest_equal_array_of_char take unsigned int lengths.

        * ktest_equal.h: Add len_unsigned_array macro for a prototype of a
        function taking an unsigned int
        count. ktest_equal_array_of_octet() is one of
        these. ktest_equal_array_of_char() takes an unsigned int as
        length.

        * trval.c (trval): Unsigned/signed int fix.

        * utility.c (asn1_krb5_data_unparse): Better test for krb5_data
        length being less than zero.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12789 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/tests/asn.1/utility.c
@@ -15,7 +15,7 @@ asn1_error_code asn1_krb5_data_unparse(code, s)
     *s = (char*)calloc(strlen("<NULL>")+1, sizeof(char));
     if(*s == NULL) return ENOMEM;
     strcpy(*s,"<NULL>");
-  }else if(code->data == NULL || code->length <= 0){
+  }else if(code->data == NULL || ((int) code->length) <= 0){
     *s = (char*)calloc(strlen("<EMPTY>")+1, sizeof(char));
     if(*s==NULL) return ENOMEM;
     strcpy(*s,"<EMPTY>");
